Add FindBugs plugin Fix all bugs detected by FindBugs Add assembly plugin Add possibility to turn on/turn off emission and absorption

+++ Filter/src/main/java/ru/nsu/fit/g14201/chirikhin/filter/model/BlurMatrix.java
@@ -17,7 +17,7 @@ public class BlurMatrix extends SquareMatrix {
                 uc = u - (size - 1) / 2;
                 vc = v - (size - 1) / 2;
 
-                g = Math.exp(-(uc * uc + vc * vc) / (2 * sigma * sigma));
+                g = Math.exp(((double) -(uc * uc + vc * vc)) / (double) (2 * sigma * sigma));
 
                 sum += g;
                 matrix[u][v] = g;
